# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyPkStore API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shop/me:
    get:
      operationId: getMyShop
      tags: [shop]
      summary: Get own shop profile
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    patch:
      operationId: updateMyShop
      tags: [shop]
      summary: Update shop settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /shop/products:
    get:
      operationId: listMyProducts
      tags: [shop]
      summary: List own products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [shop]
      summary: Add a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /shop/products/{id}:
    patch:
      operationId: updateProduct
      tags: [shop]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [shop]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /shop/orders/{id}:
    delete:
      operationId: deleteOrder
      tags: [shop]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [shop]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /stores/{slug}:
    get:
      operationId: getPublicStore
      tags: [store]
      summary: Get public store info
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStore"
        "404":
          description: Store not found
  /stores/{slug}/products:
    get:
      operationId: listPublicProducts
      tags: [store]
      summary: List products for a public store
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /stores/{slug}/orders:
    post:
      operationId: submitOrder
      tags: [store]
      summary: Customer submits a WhatsApp order
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
components:
  schemas:
...
    Shop:
      type: object
      required: [id, ownerName, shopName, slug, plan, status, whatsapp, createdAt]
      properties:
        id:
          type: integer
        ownerName:
          type: string
        shopName:
          type: string
        slug:
          type: string
        plan:
          type: string
        status:
          type: string
        whatsapp:
          type: string
        tagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        brandColor:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ShopSettings:
      type: object
      properties:
        shopName:
          type: string
        tagline:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        brandColor:
          type: string
        footerText:
          type: string
        footerAddress:
          type: string
        footerPhone:
          type: string
...
    Product:
      type: object
      required: [id, shopId, name, price, stock, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        price:
          type: integer
        stock:
          type: integer
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: integer
        stock:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
        stock:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
    Order:
      type: object
      required: [id, shopId, customerName, productName, total, status, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        productName:
          type: string
        note:
          type: ["string", "null"]
        total:
          type: integer
        status:
          type: string
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, productName]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        productName:
          type: string
        note:
          type: string
        total:
          type: integer
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
...
    PublicStore:
      type: object
      required: [id, shopName, slug, whatsapp]
      properties:
        id:
          type: integer
        shopName:
          type: string
        slug:
          type: string
        whatsapp:
          type: string
        tagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        brandColor:
          type: ["string", "null"]
        footerText:
          type: ["string", "null"]
        footerAddress:
          type: ["string", "null"]
        footerPhone:
          type: ["string", "null"]