# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  shopId: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  address?: string | null;
  productName: string;
  /** @nullable */
  note?: string | null;
  total: number;
  status: string;
  createdAt: string;
}