# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  customerName: string;
  customerPhone?: string;
  address?: string;
  productName: string;
  note?: string;
  total?: number;
}