# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  shopId: number;
  name: string;
  price: number;
  stock: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}