# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  price: number;
  stock?: number;
  category?: string;
  imageUrl?: string;
}